Backend/requirements.txt
muselsl
pyserial
tk
requests
pylsl
fastapi
uvicorn
